cambiados atributos de calificacion a enteros

+++ db/schema.rb
@@ -22,14 +22,14 @@ ActiveRecord::Schema.define(:version => 20111026031837) do
   end
 
   create_table "calificaciones", :force => true do |t|
-    t.float    "puntualidad",           :default => 0.0
-    t.float    "amigable",              :default => 0.0
-    t.float    "conocimiento",          :default => 0.0
-    t.float    "claridad",              :default => 0.0
-    t.float    "flexibilidad",          :default => 0.0
+    t.integer  "puntualidad",           :default => 0
+    t.integer  "amigable",              :default => 0
+    t.integer  "conocimiento",          :default => 0
+    t.integer  "claridad",              :default => 0
+    t.integer  "flexibilidad",          :default => 0
     t.float    "promedio"
     t.text     "critica"
-    t.float    "facilidad",             :default => 0.0
+    t.integer  "facilidad",             :default => 0
     t.float    "calificacion_obtenida"
     t.integer  "usuario_id"
     t.integer  "curso_profesor_id"
